This project is a simple calculator that is built using HTML, CSS, and JavaScript.
#Purpose of the project
The purpose of this project is to demonstrate how to build a basic calculator using the web technologies of HTML, CSS, and JavaScript. It can be used as a starting point for learning how to build more complex web applications.

#Features and functionalities
The calculator can perform basic arithmetic operations such as addition, subtraction, multiplication, and division.
It has a clear button to reset the calculator.
It has a decimal button to enter decimal points.
It displays the current equation and the result.

#Technologies used
HTML was used to create the structure and layout of the calculator.
CSS was used to style the calculator and make it visually appealing.
JavaScript was used to provide the calculator's functionality and to make it interactive.